* Keep a cache listing all, including failed, torrent loads and the mtimes of those files. This is used to ensure that e.g. the watch dir does not try to load bad/duplicate torrents, in addition to optimizing the pruning of those already loaded.
git-svn-id: svn://rakshasa.no/libtorrent/trunk/rtorrent@1024 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/utils/directory.h
@@ -48,7 +48,7 @@ struct directory_entry {
 
   // The name and types should match POSIX.
   uint32_t            d_fileno;
-  uint16_t            d_reclen;
+  uint32_t            d_reclen;
   uint8_t             d_type;
 
   std::string         d_name;
